+++ assets/editor/index.asset.php
@@ -1,1 +1,1 @@
-<?php return array('dependencies' => array('react', 'react-jsx-runtime', 'wp-api-fetch', 'wp-data', 'wp-edit-post', 'wp-element', 'wp-plugins'), 'version' => '953bec7e9bcbca29f364');
+<?php return array('dependencies' => array('react', 'react-jsx-runtime', 'wp-api-fetch', 'wp-data', 'wp-edit-post', 'wp-element', 'wp-plugins'), 'version' => 'fc5ab25b71d661daa423');